Initial Commit
Completed a working implementation of a text adventure game that allows the player to move between multiple rooms.  Now, I need to refactor the Room class to use maps instead of arrays. I also started writing implementations for the Door, TrapDoor, Person, and Dialogue classes, though these are not implemented in the main class yet.